requirements.txt
bcrypt==1.1.1
docker-py==1.1.0
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-Login==0.2.11
Flask-RESTful==0.3.2
Flask-SQLAlchemy==2.0
pytest
requests==2.6.0
python3-pika==0.9.14
